+++ backend/main/resources/resena.py
@@ -2,25 +2,31 @@ from flask_restful import Resource
 from flask import request,jsonify
 from ..models.init import ResenaModel
 from .. import db
-
-RESENAS = {
-    1: {'id_usuario': 1, 'id_comida': 2, 'calificacion': 5, 'comentario': 'Muy rica pizza!'},
-    2: {'id_usuario': 3, 'id_comida': 1, 'calificacion': 4, 'comentario': 'Hamburguesa sabrosa pero pequeña.'}
-}
+from flask_jwt_extended import jwt_required,get_jwt_identity,get_jwt
+from main.auth.decorators import role_required
 
 class Resena(Resource):
+    @jwt_required(optional=True)
     def get(self, id):
         resena=db.session.query(ResenaModel).get_or_404(id)
         return resena.to_json()
 
+    @role_required(roles=["ADMIN",'CLIENTE'])
     def delete(self, id):
         resena=db.session.query(ResenaModel).get_or_404(id)
+        rol=get_jwt().get('rol')
+        if rol=='CLIENTE' and resena.id_usuario!=get_jwt_identity():
+            return 'No tiene permiso para eliminar esta reseña',403
         db.session.delete(resena)
         db.session.commit()
         return resena.to_json(), 204
     
+    @role_required(roles=["ADMIN",'CLIENTE'])
     def put(self,id):
         resena=db.session.query(ResenaModel).get_or_404(id)
+        rol=get_jwt().get('rol')
+        if rol=='CLIENTE' and resena.id_usuario!=get_jwt_identity():
+            return 'No tiene permiso para eliminar esta reseña',403
         data=request.get_json().items()
         for key,value in data:
             setattr(resena,key,value)
@@ -29,6 +35,7 @@ class Resena(Resource):
         return resena.to_json(),201
 
 class Resenas(Resource):
+    @jwt_required(optional=True)
     def get(self):
         page=1
         per_page=5
@@ -47,6 +54,7 @@ class Resenas(Resource):
                        'pages':resenas.pages,
                        'per_page':page})
     
+    @role_required(roles=['ADMIN','CLIENTE'])
     def post(self):
         new_resena=ResenaModel.from_json(request.get_json())
         db.session.add(new_resena)
